refactor(💡): Refactor components folders to fix next build. + Adds custom pages extensions to next.config.js.
