+++ next.config.js
@@ -3,7 +3,8 @@ const nextConfig = {
   reactStrictMode: true,
   compiler: {
     styledComponents: true
-  }
+  },
+  pageExtensions: ['jsx', 'tsx']
 };
 
 module.exports = nextConfig;
