+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 // Rendered everytime the app is loaded
 import { AppProps } from 'next/app';
 import GlobalStyle from '../styles/global';
-import { Header } from './components/Header';
+import { Header } from '../_components/Header';
 
 function MyApp({ Component, pageProps }: AppProps): JSX.Element {
   return (
